requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
azure-common==1.1.28
azure-core==1.38.1
azure-search-documents==11.6.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.129.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.13.0
openai==2.21.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.7.1
python-dotenv==1.2.1
python-multipart==0.0.22
requests==2.32.5
sniffio==1.3.1
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
